insurance-mcp/pyproject.toml
[project]
name = "insurance-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "certifi==2025.10.5",
    "click==8.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "mcp[cli]==1.19.0",
    "openai>=2.6.1",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pymongo==4.10.1",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "referencing==0.37.0",
    "requirements-py>=1.0.1",
    "rpds-py==0.28.0",
    "sniffio==1.3.1",
    "sse-starlette==3.0.2",
    "starlette==0.48.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.38.0",
]
